refactor: handleHTTPError 모듈 분리
- handleHTTPError 함수를 utils/error.ts로 이동
- 상수 및 타입 분리
